requirements.txt
Django==3.2.16
azure-identity==1.10.0
azure-storage-blob==12.13.1